VE-2755: Remove UtilBase trait in favor of object singleton.
This is because `Util`, as the name suggests, is a utility, like a 'simple library' kind of API, not an SPI (Service Provider Interface) kind of API.
So for that, is best to *not* use inheritance (see [ELE office hours video on traits](https://drive.google.com/file/d/1XC06TPu71CYr1BSLDKb1ifZPviGD0eQI/view), see book Practical API Design, Yaroslav Tulach).
If injection will be needed, it is best to use it within other points of extension, perhaps where Util depends on, not in the very methods of Util.

This change is simply about:
- copy over the content of UtilBase.scala into Util.scala
- rename UtilBase class name to Util
- remove UtilBase.scala

Other changes were about:
- remove `AsyncToSync`. This is towards removing use of method `DEPRECATED_convertToSyncReq()` where possible
- removing inheritance of `TestUtil` from `UtilBase` (it was not used)
- calling methods of `TestUtil` directly where needed (TestUtil being also a 'simple library' kind of API)